// src/presentation/store/minicart/types.ts
export type CartState = {
  cartBFF: Cart ;
  cartId: string ;
  error: string;
  loading: boolean;
};